Create README.md
FriendFinder
This is homework 11 for the Northwestern Coding Boot Camp.

This application is essentially a non-creepy friend pairing service. Here is the link to the app on heroku.

Setup
First clone this repo to your local directory.

Next, navigate to your repo in your command terminal, and enter:

npm install

This will ensure you have all the necessary NPM packages installed to run this generator.

Feel free to load your own starter data into the friends.js file, or use the ones already in there. It's ok, I don't know them either.

Overview
The home page of the app serves the purposes of telling you about the app.

Navigating to the survey page will present you with a 12 item questionnaire: name, photo link, and 10 questions to determine your fate.

All questions are required.

Clicking submit will generate your match and display it to you in a pop modal.